+++ src/Concesionario/Concesionario/Cliente.cs
@@ -4,7 +4,7 @@ using CsvHelper.Configuration;
 
 public class Cliente
 {
-    public int Cedula { get; set; }
+    public long Cedula { get; set; }
     public string Nombre { get; set; }
     public string Telefono { get; set; }
     public string Direccion { get; set; }
