+++ src/Concesionario/Concesionario/Factura.cs
@@ -1,12 +1,12 @@
 public class Factura
 {
-    public int Id { get; set; }
+    public long Id { get; set; }
     public DateTime Fecha { get; set; }
     public Cliente Cliente { get; set; }
     public List<Vehiculo> Vehiculos { get; set; }
     public decimal Total { get; set; }
 
-    public Factura(int id, DateTime fecha, Cliente cliente, List<Vehiculo> vehiculos, decimal total)
+    public Factura(long id, DateTime fecha, Cliente cliente, List<Vehiculo> vehiculos, decimal total)
     {
         Id = id;
         Fecha = fecha;
