+++ src/Concesionario/Concesionario/Vehiculo.cs
@@ -1,7 +1,7 @@
 
 public abstract class Vehiculo : IVendible
 {
-    public int Id { get; set; }
+    public long Id { get; set; }
     public string Marca { get; set; }
     public string Modelo { get; set; }
     public string Color { get; set; }
@@ -10,7 +10,7 @@ public abstract class Vehiculo : IVendible
     public int Cilindraje { get; set; }
     public bool Vendido { get; set; }
 
-    public Vehiculo(int id, string marca, string modelo, string color, string placa, decimal precio, int cilindraje, bool vendido = false)
+    public Vehiculo(long id, string marca, string modelo, string color, string placa, decimal precio, int cilindraje, bool vendido = false)
     {
         Id = id;
         Marca = marca;
